requirements.txt
pycryptodome==3.19.0
gostcrypto==1.2.5
pytest==7.4.3
pytest-cov==4.1.0 